feat: Add LED_STATUS_ERROR_BUSY to LED animation error codes
The code changes in `led_animation.h` add a new error code `LED_STATUS_ERROR_BUSY` to the LED animation error codes. This error code is used to indicate that the LED is currently busy and cannot perform the requested operation. This change improves the error handling and provides more specific feedback to the user.